+++ tests/providers/test_provider.py
@@ -1,0 +1,132 @@
+# Copyright 2021 Canonical Ltd.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# For further info, check https://github.com/canonical/charmcraft
+
+from unittest import mock
+
+import pytest
+
+from charmcraft import providers
+from charmcraft.config import Base
+
+
+@pytest.fixture(autouse=True)
+def mock_get_host_architecture():
+    with mock.patch(
+        "charmcraft.providers._provider.get_host_architecture", return_value="host-arch"
+    ) as mock_arch:
+        yield mock_arch
+
+
+def test_get_command_environment_minimal(monkeypatch):
+    monkeypatch.setenv("IGNORE_ME", "or-im-failing")
+    monkeypatch.setenv("PATH", "not-using-host-path")
+    provider = providers.LXDProvider()
+
+    env = provider.get_command_environment()
+
+    assert env == {
+        "CHARMCRAFT_MANAGED_MODE": "1",
+        "PATH": "/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/snap/bin",
+    }
+
+
+def test_get_command_environment_all_opts(monkeypatch):
+    monkeypatch.setenv("IGNORE_ME", "or-im-failing")
+    monkeypatch.setenv("PATH", "not-using-host-path")
+    monkeypatch.setenv("http_proxy", "test-http-proxy")
+    monkeypatch.setenv("https_proxy", "test-https-proxy")
+    monkeypatch.setenv("no_proxy", "test-no-proxy")
+    provider = providers.LXDProvider()
+
+    env = provider.get_command_environment()
+
+    assert env == {
+        "CHARMCRAFT_MANAGED_MODE": "1",
+        "PATH": "/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/snap/bin",
+        "http_proxy": "test-http-proxy",
+        "https_proxy": "test-https-proxy",
+        "no_proxy": "test-no-proxy",
+    }
+
+
+@pytest.mark.parametrize(
+    "bases_index,build_on_index,project_name,target_arch,expected",
+    [
+        (0, 0, "mycharm", "test-arch1", "charmcraft-mycharm-{inode}-0-0-test-arch1"),
+        (
+            1,
+            2,
+            "my-other-charm",
+            "test-arch2",
+            "charmcraft-my-other-charm-{inode}-1-2-test-arch2",
+        ),
+    ],
+)
+def test_get_instance_name(
+    bases_index, build_on_index, project_name, target_arch, expected, mock_path
+):
+    provider = providers.LXDProvider()
+
+    assert (
+        provider.get_instance_name(
+            bases_index=bases_index,
+            build_on_index=build_on_index,
+            project_name=project_name,
+            project_path=mock_path,
+            target_arch=target_arch,
+        )
+        == expected.format(inode="445566")
+    )
+
+
+@pytest.mark.parametrize(
+    "name,channel,architectures,expected_valid,expected_reason",
+    [
+        ("ubuntu", "18.04", ["host-arch"], True, None),
+        ("ubuntu", "20.04", ["host-arch"], True, None),
+        ("ubuntu", "20.04", ["extra-arch", "host-arch"], True, None),
+        (
+            "not-ubuntu",
+            "20.04",
+            ["host-arch"],
+            False,
+            "name 'not-ubuntu' is not yet supported (must be 'ubuntu')",
+        ),
+        (
+            "ubuntu",
+            "10.04",
+            ["host-arch"],
+            False,
+            "channel '10.04' is not yet supported (must be '18.04' or '20.04')",
+        ),
+        (
+            "ubuntu",
+            "20.04",
+            ["other-arch"],
+            False,
+            "host architecture 'host-arch' not in base architectures ['other-arch']",
+        ),
+    ],
+)
+def test_is_base_available(
+    mock_get_host_architecture, name, channel, architectures, expected_valid, expected_reason
+):
+    base = Base(name=name, channel=channel, architectures=architectures)
+    provider = providers.LXDProvider()
+
+    valid, reason = provider.is_base_available(base)
+
+    assert (valid, reason) == (expected_valid, expected_reason)
